AEG-414 - only add wormhole error handling for custom resolvers
We may want to make this configurable at some point, until now we've
been wrapping _every_ middleware in Wormhole and that's not free. For
large response payloads you can have 1,000 or more nodes in your graph
and most of them are just using the default `MapGet` middleware that
shouldn't really ever throw an error. If we skip the Wormhole wrapping
for those middleware things will go a lot faster.